HS-849 Split node overlay
